api/requirements.txt
fastapi>=0.78.0
python-decouple>=3.6
python-socketio>=5.7.2
redis>=4.3.5 # required for socketio